[lnt] Expose ability to run under qemu.
Adds a --qemu-user-mode=VALUE option that sets USER_EMULATION_MODE to '1' and
RUNUNDER to the given value. It also records the use of qemu in report.json.

If qemu's default options and environment are sufficient then the option value
should be a qemu binary. If you need special options or environment variables,
then the option value should be a rununder script. In future, there will
hopefully be a way to generate a trivial rununder script to add options such as
'-cpu foo'.

llvm-svn: 203115