Improved regression detection with min of diffs
This change reduces false positives in LNT's regression detection
algorithm by having it use more past data.  Instead of calculating the
delta as the difference of min(current)-min(prev), we use the smallest
difference between min(current) and all prevs.  When the data has noise
or is multimodal this reduces our chances of flagging a false positive
regression.

llvm-svn: 238965