Add support for storing hash of test binaries.
This commit adds support for storing hashes of the test binaries as a new
sample type. The idea is that when the hash function of the same test is the
same as on a previous run, the analyses can assume that the binary didn't
change in a meaningful way.

None of the analyses in LNT have been adapted yet to take advantage of this
extra information.

The hash value is stored as a new sample type.  Since the hash value in many
cases should not be handled the same as other sample types, like execution time
or compilation time, a distinction needs to be made. This patch changes
terminology in LNT and calls sample types that have a meaningful order, like
execution time and compilation time, "metric fields". The hash function of a
binary - as one hash value cannot be said to be "better" than another hash
value - is not considered a "metric field".

Care has been taken to make sure LNT can keep on working with test-suite
versions that don't produce the hash values yet, which was introduced in
r246035.

Differential Revision: http://reviews.llvm.org/D11231

llvm-svn: 249026