[html] Move deprecated align/width/bgcolor attributes into the style attribute. NFC.
The align, width and bgcolor attributes from tags <td> and <th> are obsolete
(refer to http://www.w3.org/TR/html-markup/td.html#td-constraints and
http://www.w3.org/TR/html-markup/th.html#th-constraints). The CSS has
to be used instead. CSS is however not an option in our case because
some mail clients (like Gmail) do not accept them, so we have to pass
those deprecated attributes in the style attribute.

We are now constructing the style attribute from a part coming from the
python script (with the styles dict) plus a part from the template
itself. Should the relevant styles.attr be undefined (or empty), an
empty string will be produced; there will be an extra semicolon, but
this's valid syntax.

llvm-svn: 253375