Add a manual regression creation link to the graphs
Sometimes a change has not been detected, it is nice to be able to
create a regression anyways.  To do this we take the runID and url
which have been exposed into the graphs, and create a server endpoint
that works a lot like fieldchange generation processes. It will always
produce a regression though, even if the changes are not that
interesting.

llvm-svn: 256297