[html] Advertize the charset actually being used. NFC.
Summary:
We currently advertize the ISO-8859-1 charset is used when we
actually use UTF-8. Let's be consistent an advertize the truth.

This is the last patch in a serie which has brought the main, machine, run and
daily_report pages into clean html (according to https://validator.w3.org/nu).

Reviewers: cmatthews, ddunbar

Subscribers: llvm-commits

Differential Revision: http://reviews.llvm.org/D14830

llvm-svn: 258674