[lnt] Register signal handlers on the main thread
We've seen tracebacks like the one below on our LNT server, indicating
that previously, the code that adds the signal handler could be executed
by a thread that's not the main thread. Fix this by installing the
signal handler at module load time, which is assumed to always run in
the main thread.

Patch by Kristof Beyls, reviewed by Chris Matthews

http://reviews.llvm.org/D17508

llvm-svn: 263416