[profile] Add the initial scaffold for the LNT profiling infrastructure
This adds the API that will be used to query profiles in LNT's web stack along with a dumb implementation.

The dumb implementation serves as a way to easily import profile data (from whatever profile source). The auto-upgrade functionality in the profile API will then allow that to be upgraded to a more complex, more compact profile representation to be added later.

llvm-svn: 263559