[profile] Add perf and cPerf, for importing Linux Perf profiles
cPerf is a C++ python extension that reads perf.data sample streams and aggregates the data (including running objdump) and producing a ProfileV1 object.

The module is marked "optional" in setup.py, which means that if it fails to build it shouldn't fail setup (and perf.py allows module import to fail). However the version of distutils I have here on a mac doesn't understand that option, so hopefully it won't fail to build...

Initially I implemented this as a simple regex wrapper around 'perf annotate' that grepped the counter values and objdump output from perf's textual output.

Writing this in C gained me about a 100x speedup. In fact, this C implementation is around 6x as fast as perf itself, probably because perf does more magic.

An optimization here is that we ignore any binary with < 1% of the samples. This in practice means we don't waste time nm'ing and objdumping things like "perf_4.0" and "/bin/bash" or most of the dynamic libraries. This also gives a rather large speedup.

llvm-svn: 263560