[profile] Add an optimized on-disk representation
ProfileV2 is a profile data representation designed to keep the
profile data on-disk as small as possible while still maintaining good
access speed, specifically to avoid having to load the entire file to
discover just the index (functions and top level counters).

The format:
  * Is a binary format consisting of an index and several sections.
  * Consists of only two datatypes:
    * Strings (newline terminated)
    * Positive integers (ULEB encoded)
  * Some sections are expected to be BZ2 compressed.

The TextPool section has the ability to be shared across multiple profiles to
take advantage of inter-run redundancy (the image very rarely changes substantially).
This pooling ability is not yet implemented but the appropriate scaffolding is in
place.

The ProfileV2 format gives a ~3x size improvement over the ProfileV1 (which is also
compressed) - meaning a ProfileV2 is roughly 1/3 the size of ProfileV1. With text
pooling, ProfileV2s can be even smaller.

llvm-svn: 263562