[profile] Add an admin page to view the current profile stats
This page shows graphs of the total profile size (in MB) and a histogram of the age distribution of the profiles in the cache.

This allows for monitoring of the profile cache makeup. It takes its input from two JSON files that are not yet populated but will be soon.

llvm-svn: 263880