[ui] Use Bootstrap CSS styling for run report tables where possible
The run reports are rendered in two ways: for email and for web. Email doesn't do CSS very well, so we were styling the tables ourselves.

However, on web we have a real theme and the manually styled table doesn't fit it.

Not only does this make the table fit the rest of the page better, it makes it more suitable as a container for profile view links, which will be added in a followup commit.

llvm-svn: 263925