[profile] Implement support for profiling shared objects
Shared objects are relocated, which means we need to adjust the PC value by subtracting the map base. However we need to determine the difference between a shared object (DYN) and an executable (EXEC) to do this (because executables' PCs should not be adjusted). We therefore load in part of the ELF header and test for e_type == ET_DYN.

llvm-svn: 264042