[ui] Add a mechanism for selecting different compare_to and baseline runs
In the v4_runs page, we compute a run to compare to and a baseline run. It's possible for the user to pick a different run to compare to by choosing from the past 5 previous runs, but if the user wants to compare to a specific run, he has to go find that run, look up its run ID, and manually edit the URL, adding "?compare_to=ID".

A similar procedure is required for changing the baseline.

This patch adds an easier mechanism. Hovering over the table rows for "previous" or "baseline" reveals an edit icon - clicking on that launches a popover that contains a search box.

The search mechanism (runTypeahead) was taken from the profiling infrastructure and has been found a new home in lnt_run.js.

llvm-svn: 264723