Server side blacklist for changes
There are some changes we don't care about.  For instance, changes in
Unittests and the known to be evil benchmark ks.  This implementes a
regex based blacklist for LNT change detection.  A blacklist path can be
added to your config file which will point to a file with regex per
line, of things to ignore.  When we do change detection, any changes
that match those regexes will be dropped.  A default blacklist is not
provided, so nothing is filtered unless you set it up.

llvm-svn: 267787