Avoid using Flask 0.11
It seems some parts of LNT depend on functionality in Flask that was
removed in Flask 0.11. Indicate in setup.py that we need a version of
Flask before 0.11.

Change inspired by documentation at
https://caremad.io/2013/07/setup-vs-requirement/ and
http://python-packaging-user-guide.readthedocs.io/en/latest/requirements/.

The backtrace seen on bots after Flask 0.11 was released and LNT
started picking that version when being installed ends in:

  File "test/lnt/lnt/server/reporting/runs.py", line 7, in <module>
    import lnt.server.ui.app
  File "test/lnt/lnt/server/ui/app.py", line 24, in <module>
    import lnt.server.ui.views
  File "test/lnt/lnt/server/ui/views.py", line 29, in <module>
    from lnt.server.ui.decorators import frontend, db_route, v4_route
  File "test/lnt/lnt/server/ui/decorators.py", line 6, in <module>
    frontend = flask.Module(__name__)
AttributeError: 'module' object has no attribute 'Module'

llvm-svn: 271274