Builds are not clean between repeated runs
Motivation:
When doing a compile multiple times in a row only the first build is actually happening.
All subsequent runs to build do nothing as make detects that none of the files have changed.

What changed:
This deletes the entire build directory before each compile happens, thereby ensuring that each run has a clean build.
Since the build directory is deleted before the compile, the "last" build directory is kept. In case there is a failure this build directory can be used to look at what happened.

To delete the build directory I am using the -p flag of runN to perform the cleaning before the timing run.
Differential Revision: https://reviews.llvm.org/D28882

llvm-svn: 292527