Add global, user selectable, baselines to LNT
This feature adds named baselines to LNT, and then switches
between them.  The motivation for this is to allow us to do better
tracking for many baselines, for example clang-3.9, and clang-4.0 at
the same time. Currently LNT supports one baseline per database, but
it is desirable to be able to compare with many known points in history.
The new baselines are named, to make it simpler to know what the meaning
of the baseline is.

The Order page has been updated to allow an order to be named and
promoted to a baseline.  Once promoted, that baseline will be shown in
a new global baseline dropdown.  Users can select a baseline from the
dropdown, which is stored in their session and will be used in all pages.

This commit changes the matrix page to use the new baselines. Future
commits will change other pages to use the new baselines, when the old
baseline data is no longer used I will remove the old config file
baseline.

llvm-svn: 293985