Run configure earlier in runtest test-suite [NFC]
This is in preparation for follow-on work where we'll extract the
compiler used from the CMakeCache rather than from the --cc command line
option. For that to work, however, the configure step has to be run
before the first time self._get_cc_info() gets called.

This also simplifies the need-to-configure logic a bit.

Differential Revision: https://reviews.llvm.org/D29821

llvm-svn: 294935