Use suite and db specific keys for baseline selection
Baselines are stored at the test-suite and db level, but the session
key we used did not encode that.  That would mean if selected a baseline
then moved to a different DB, a differnt baseline would be randomly
selected, or LNT would crash.

Now we encode the DB name and suite in the session key, so we know which
baseline is which.

llvm-svn: 299253