[cPerf] Timestamp MMAP and MMAP2 events
Events in the perf event stream may not be in order. We don't generally
care if samples are out of order because we don't use their timestamp
information. However we *do* care if a sample appears before the MMAP
event that defined which file was mapped to its PC.

MMAP events are timestamped. Change the data structure for PC->mmap
lookups so that we can query the state of the address space mappings at
a particular time. This isn't mega efficient, but we don't expect many
MMAPs.

Patch by James Molloy.

llvm-svn: 300680