Fix LNT’s js to corrrectly determine base URL when server has a prefix
There was a number of errors in the JS leading to a invalid prefix being
generated at runtime. That made some of the URLs and ajax requests
invalid. The prefix was being overwritten, and had the wrong value.

Now we use a global determined by Flask (which should know these things)
and update the urls to not overwrite each other.

llvm-svn: 302269