async_ops.py: only set up worker processes once
This was obviously the intended behaviour but since WORKERS never got set to anything we ended up creating new processes for every async process. This is all well and good until one async process outlasts a newer one, in which case the log buffer object it writes to has been destroyed in the parent and multiprocessing madness ensues.

Fixes PR33067.

llvm-svn: 303244