Add postgres_wrapper.sh to enable tests without server
Add a wrapper script that creates an ad-hoc postgres instance so you can
run the unittests without setting up a real server on your machine.

You still need to have postgres installed and need to enable the test
with `lit -Dpostgres=1`, but at least there is not need to setup and
start a server anymore.

llvm-svn: 306326