Add a delete method with auth to the runs API
Add a token based auth system to the API.  Add a delete method to the
runs endpoint for authorized users.  This will make it easier for non
admins to delete their data.  Hopefully we can get a better auth system
in place, but for now it is single key per instance.

llvm-svn: 306901