Move some v4db methds down to TestSuiteDB
This moves getNum{Machines|Runs|Samples|Tests}() and
importDataFromDict() from v4db into TestSuiteDB so they only operate on
a single testsuite.

They are only ever used to compute the delta when inserting into
a single test-suite anyway, so we are fine with the simpler
per-testsuite version.

It also fixes wrong numbers getting reported when submitting to a
testsuite for the first time after a server restart. getNumXXX() would
not see the testsuite at first as it wasn't loaded yet but after loading
the suite and inserting it would see all the old data and report bogus
numbers.

llvm-svn: 307415