Use yaml schema mechanism to describe compile suite
This re-applies r307562 now that the metatables are upgraded in sync
with the yaml so the fieldchange mechanism should work again and migrate
cleanly. It also comes with a more robust migration script.

This converts the existing compile suite schema which was previously
configured via the TestSuiteXXX tables in lnt/server/db/migrations to
use the new yaml schema mechanism instead.

- Adds schemas/compile.yaml with the test suite definition.
- Adds a migration that removes the compile definitions in the
  TestSuiteXXX meta tables. It also drops the compile tables if they are
  empty (and otherwise leaves them untouched).
- This means by default the compile suite will not be visible anymore.
  Users using the 'compile' suite need to copy or symlink
  schemas/compile.yaml into $LNTINSTANCEPATH/schemas/compile.yaml

llvm-svn: 307963