Fix missing "Produced by" field on Run causing UI crash.
When a given run contains a "producer" value (usually an URL pointing to build
systems), the run details page can crash when displaying this run in cases when
the previous or baseline runs do not contain a valid value for the "producer"
field.

It is possible to see the following error message:

...
File "/sandbox/lnt/server/ui/filters.py", line 47, in filter_producerAsHTML
  if not producer:
jinja2.exceptions.UndefinedError: 'dict object' has no attribute 'producer'

This change prevents the UI to crash when "producer" is not found. It also adds
a test change to capture this situation.

Patch by Leandro Nunes.
Differential Revision: https://reviews.llvm.org/D35001

llvm-svn: 308019