api: When removing a machine, delete runs in chunks.
For machines with hundreds of runs we run into timeouts and OOM
situations when sqlalchemy tries to delete everything in a single atomic
operation.

Change the code to delete the runs of the machine in chunks and report
status updates back to avoid timeout and give a sense of progress to the
user.

llvm-svn: 308410