LNT API run roundtrips
Allow LNT runs to be downloaded from the API in a format that they can
be inserted back into the same or other LNT instances.  This can be used
to move runs between LNT instances, or modify them.

I make some changes to the REST API to make the format closer to what we
expect, then ignore some fields (like ID) when importing.

llvm-svn: 308565