Rework machine creation strategy
Currently when submitting a run and the machine data does not match the
previous data, LNT creates a new machine with the same name (but
different id). This was often very confusing to users.

This changes the strategy to reject the submission if the data does not
match the previous data unless the `--update-machine` flag (or the
update_machine post parameters, etc.) is set in which case the new data
overrides the previous machine data.

Adding previously unset keys to the machine will not lead to a rejection
either way. Leaving out previously set keys is fine (but will not remove
the keys).

This new strategy will result in machine names being unique except for
the case of older entries in the database before this change.

Differential Revision: https://reviews.llvm.org/D35598

llvm-svn: 309247