Do not store index with SampleField object; NFC
The index field cannot be computed in the constructor of the SampleField
object; it was only initialized when the TestSuite object was created.
The problem with this approach is that when querying SampleFields from
the database, the index field will only be populated when the sqlalchemy
cache hits the objects where we previously initialized it. When using
multiple/shorter lived sqlalchemy sessions the cache won't unique the
instances.

This removes the index member from SampleField and instead provide the
TestSuiteDB.get_field_index() method that computes the index for a given
SampleField object. This helps improving LNTs sqlalchemy-session usage
in the future.

llvm-svn: 312059