Remove --commit/--commit=0 flags/settings.
This removes the possibility to not commit a run. It was easy to forget
to pass the commit flag by accident while when you wanted to check the
submission syntax you could just as well use `lnt checkformat` instead
of submitting without commit.

This also reverts most of the changes in r308256; as there is little use
in making the flag simpler to use when we can just remove it instead.

Note that the --commit=0/1 flags are still accepted (but ignore) to
accomodate old scripts. We also pass 'commit': '1' when submitting to
server so that submissions to old servers still work.

Differential Revision: https://reviews.llvm.org/D35588

llvm-svn: 312068