async_ops: Do not dispose pre-fork DB connections
It turns out that the hack in async_ops that disposes the DB connections
present before forking the new process is actively harmful (at least for
postgres/psycopg2 setups) and appears to disrupt the connections of the
original process. The recent session cleanup seems to have made this
hack obsolete so remove it.

llvm-svn: 312741