docs/Makefile: Change html target back to not copying the files around
Not everyone builds the docs into the source directory or wants to
update the copy in server/ui/static. This changes `make html` back to
only create the documentation files as it was before r255982. Instead
add a new target "copydocs" which copies the files after building them;
make 'copydocs' the default target.

llvm-svn: 317745