Use the internal shtest runner and make lit a formal dependency
The internal shtest runner is able to print command output etc, so it is
much better for debuggined test failures, especially in CI.  On my
machines ToT lit is able to run all the tests with the internal runner.

Since a newish version of lit is needed (it is for test running as well)
I will add it as a formal dependency.

llvm-svn: 331289