[LNT] In command-line tool allow to use relative path for test suite location.
When test command `test-suite` is executed, it runs inside the sandbox and uses
test suite path as-is. So the path relative to the working directory is not
found when it is relative to the sandbox directory. Fix by converting relative
path to absolute path.

Test command `nt` works with relative paths correctly because it runs tests in
a different way.

Reviewers: cmatthews, kristof.beyls, MatzeB

Reviewed By: cmatthews

Subscribers: dexonsmith, llvm-commits

Differential Revision: https://reviews.llvm.org/D50916

llvm-svn: 340107