Scaleability improvements: don't expire objects and use better query
When we commit we expire all objects in the session, meaning the will be
lazy loaded again next time they are accessed.  Move the commit of all
found regressions to the end of processing, so the caches stay valid for
the entire life of analysis.

Also, use FieldChange directly for analysis, instead of RegressionInd,
we were mostly only accessing the change field anyways.

llvm-svn: 349578