[LNT] Python 3 support: adapt to zip returning an iterator
Summary:
zip() returns a list in Python 2 but an iterator in Python 3.
Fortunately, the only use is in the return statement in pairs() and the
only call site of pairs() iterates over the result. This commit thus
adds a docstrings for pairs to make it clear callers can only rely on
the result being iterable. It also rename the parameter to not conflict
with the list builtin function.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls

Reviewed By: hubert.reinterpretcast

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D67819

llvm-svn: 372826