[LNT] Python 3 support: Do not encode report for stdout
Summary:
Daily report and run comparison are both output to stdout and sent by
email as attachment. Under Python 3's strict separation between text and
binary data the former needs to be done as a string while the latter
needs to be an encoded text, i.e. as binary data. This commit implements
that separation.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: PrzemekWirkus

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68797