[LNT] Python 3 support: fix printing of exceptions
Summary:
A number of SQLAlchemy exception are printed by using their message
attribute. However PEP-352 deprecates this attribute from Python 2.6
onwards and drop it for Python 3 and later versions.

This commit replaces printing the message attribute by printing the
exception itself, relying on the __str__ method. This is actually
equivalent for exceptions with only a single argument (as required for
all exceptions on Python 3) as can be seen by the message property
getter in the aforementioned PEP.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: PrzemekWirkus

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68798