[LNT] Fix sql index name collision
Summary:
As revealed by the exception printing failure prior to the previous
commit, sqllite throws an error when creating the index for the
RegressionID field for the NT and link testsuite schemas since they have
the same name. This commit adds the testsuite schema to the index name
to avoid this collision.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: PrzemekWirkus

Subscribers: arphaman, llvm-commits

Differential Revision: https://reviews.llvm.org/D68799