[LNT] Python 3 support: subprocess IO as text
Summary:
In Python 3, subprocess stdin, stdout and stderr are in binary mode by
default. Trying to print or parse those as text as is done thoughought
in LNT then causes byte/str errors. This commits sets the
universal_newlines parameter to True when invoking subprocess methods
involving input or output to request those to be in text mode. It also
adapt _check_output to work whether universal_newlines is set to True in
the caller or not. Finally, it cleans up 2 cases subprocess method call
where the output file objects are stored in a variable but not used.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: PrzemekWirkus

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68938