[LNT] Python 3 support: adapt csv reader open
Summary:
csv.reader have different expectation between Python versions regarding
the file object it reads from. On Python 2 it requires the file to have
been opened in binary mode while on Python 3 it requires text mode and
universal newlines mode enabled but forwarded untranslated to the
caller. This commit use an if to cater to both requirements.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: PrzemekWirkus

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68800