[LNT] Python 3 support: fix text/binary confusion in profile support
There is confusion between text and binary in the checkFile method of
the various profile format implementation modules. All of them compare
the first few bytes read in text mode against an expected value.
However, the header of the profile v1 contains some binary data, thus
causing an error in Python 3 when reading a profile v1 file and calling
any of the checkFile method.

This commit changes the checkFile method to read binary data and compare
against binary literals.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls

Reviewed By: hubert.reinterpretcast

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68222

llvm-svn: 373617