[LNT] Python 3 support: fix text/binary confusion in importreport
lnt importreport currently fails in Python 3 because it opens the input
and output file in binary mode. Yet, as per documentation, this tools
converts "*text based* key value pairs into a LNT json report file".
The JSON file is output using the json module which expects a file
opened in text mode so this commit changes the mode of both the input
and output file to text.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls

Reviewed By: hubert.reinterpretcast

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68224

llvm-svn: 373861