[LNT][NFC] Fix global import in function
Summary:
The _load_dependencies() function in lnt.lnttool.admin import some
modules in the global namespace by making the module name global and
then doing imports. However the language reference for global statements
forbids this:

"Names listed in a global statement must not be defined as formal
parameters or in a for loop control target, class definition, function
definition, or import statement."

This commit makes use of importlib.import_module wrapper instead of the
import statement to respect this restriction.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, PrzemekWirkus

Reviewed By: PrzemekWirkus

Subscribers: PrzemekWirkus, leandron, MatzeB, llvm-commits

Differential Revision: https://reviews.llvm.org/D68779