[LNT] Python 3 support: fix writing report to tmp file
Summary:
The import_from_string() method in lnt.util.ImportData writes a copy of
the report being imported into a temporary file. That fails under Python
3 because by default tempfile.mkstemp open in binary mode and the data
written is text.

This commit sets the text parameter of mkstemp to True to ensure the
file is opened in text mode. It also write to the file via a file object
since os.write expects binary data.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: cmatthews

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68802