[LNT] Python 3 support: sort benchmark regressing
Summary:
Test server/ui/change_processing.py makes an assumption about the order
in which benchmark are listed when a regression occurs but these are put
in a set before being iterated over for printing. Since set have no
guarantee of order, the test is fragile and fails on Python 3 due to
different implementation of Set. This commit sort the benchmark when
printing them so that the output is consistent. It also adapts the
testcase accordingly.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: PrzemekWirkus

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68803