[LNT] Python 3 support: fix report version literal
The upgrade_1_to_2 module stores a __report_version__ string literal in
the Parameter column of the Run table which is declared as binary data.
This does not work in Python 3 where strings/text and binary are
different things. This commit changes it for a binary literal instead
(introduced in Python 2.6).

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls

Reviewed By: hubert.reinterpretcast

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68105

llvm-svn: 374304