[LNT] Python 3 support: specify how to sort dict
Summary:
To be able to compare 2 lists of dictionaries, test
server/ui/test_roundtrip.py calls sorted on those list so that they are
comparable. However no comparison function is given so it fails on
Python 3 because dictionary are not comparable. This commit adds a key
parameter to sort on the name key of these dictionaries.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: cmatthews

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68920