[LNT] NFC: Fix order of globals and locals on exec
Summary:
Per https://docs.python.org/3/library/functions.html#exec, the globals
parameter comes before the locals one. Since `globals` and `locals`
refer to the same object for the call in question, we can remove
`locals`, which will cause the globals parameter to be used for both the
global and the local variables, thus keeping the same behavior.

Reviewers: cmatthews, thopre, kristof.beyls

Reviewed By: thopre

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D68903

llvm-svn: 374685