[LNT] Python 3 support: define hash for Order class
Summary:
Method v4_machine in lnt.server.ui.views stores couples of Order and Run
instances in a multidict, thus making Order the key. This requires
instances of the Order class to be hashable but hashability rules have
changed under Python 3.

As per the Python's data model documentation regarding the __hash__
method, user-defined classes have a default __hash__ method making their
instances hashable. However, under Python 3 if such a user-defined class
overrides __eq__ but not __hash__ (as is done by the class
lnt.server.db.testsuite.TestSuiteDB.Order) the __hash__ method is set to
none making it unhashable. Under Python 2 one would have to explicitely
set __hash__ to None in the class definition for such a class to be
unhashable.

This commit adds a __hash__ method for
lnt.server.db.testsuite.TestSuiteDB.Order class to allow it to be used
as key in a multidict.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: cmatthews

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D69042