[LNT] Python 3 support: run tests with UTF-8 encoding
Summary:
Click requires to be invoked with a default encoding of UTF-8. If not,
it exists the program with an error. This commit sets the LANG
environment variable for running lit tests to the system locale if it is
a UTF-8 capable one or en_US.UTF-8 as a fallback. This allows the test
to run on systems where en_US.UTF-8 is not installed (such as some
container images) but the default locale is UTF-8 capable.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, leandron, PrzemekWirkus

Reviewed By: hubert.reinterpretcast

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D69048