Remove all past crazyness with pip in setup.py
In an attempt to be DRY, we were using the requirements files to get the dependent
packages. Pip has not provided a good interface to do that, and so we have been
broken a number of times.  As pip has evolved, those files can now define dependencies
better, so they are not needed anymore.

In a standard client install (flask webserver and sqlite) we will install the packages
using standard setup.py install_requires.  For server installs pip can be used directly,
and will use the local setup.py to get the non-server requirements.  This is much more
standard.

Docs updated, but Docker change not tests, as I don't have a test environment.