Fix dynamic import of http.client
Summary:
Contrary to __import__, importlib.import_module returns the specified
package or module while __import__ returns the top-level package or
module and is thus what http would be bound to when importing
http.client. This commit fixes the import of http.client so that the
http module is bound to the http variable. Without this change LNT
crashes when receiving an HTTP response not between 200 and 400 (not
included).

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls, PrzemekWirkus

Reviewed By: cmatthews

Subscribers: llvm-commits, MatzeB, leandron, PrzemekWirkus

Differential Revision: https://reviews.llvm.org/D79183