Fix pip install line in Dockerfile
Summary:
The pip install line in Dockerfile is missing the -r switch and is
not executed from the directory where the requirement file is, both of
which make the docker image fail to build. This commit fixes those
issues and also simplify the line by using pip3 instead of
python3 -m pip.

Reviewed By: cmatthews

Differential Revision: https://reviews.llvm.org/D79238