Fix crash in psycopg2 with docker image
Summary:
Psycopg2 2.8 crashes when used with PostgreSQL 12 [1] which is now the
default version of the postgres:alpine docker image. This commit sets
the tag for the postgres image used by the LNT service to 11-alpine to
prevent this issue. Going forward the requirement for psycopg2 will need
to be bumped again to 1.3.16 or later but given how new is that release
it's better to fix the PostgreSQL version for now.

[1] https://github.com/sqlalchemy/sqlalchemy/issues/4463

Reviewed By: cmatthews

Differential Revision: https://reviews.llvm.org/D79250