Generate relative URLs instead of absolute urls for weburls.
Currently LNT generates for the web views a url by using flask, however the documentation[1] specifies that when _External is true then an absolute URL is generated.  This base url is used in links generated by the page.

This means we now emit absolute urls into the html. This is problematic as it breaks when LNT is used behind a reverse proxy.  The urls used then won't match and will be incorrect.

OK for master? Also I don't have commit rights so if OK Can someone commit it for me?

[1] https://flask.palletsprojects.com/en/1.1.x/api/#flask.url_for

Reviewed By: thopre

Differential Revision: https://reviews.llvm.org/D85425