Prevent page crash when using default comparer on Global Status page
When LNT was still Python 2 the default comparer was able to handle None
values with <, however Python 3 doesn't support this. As a result after
the Python 3 upgrade the Global Status page crashes when sorting any
values with None in them in the second column.

I believe this happens when there is no regression detected so this is a
valid state.

There are two ways to handle this, either filter out the None values or
sort them all the same. I have choosen the second approach since I
believe that is what Python 2 did.

Reviewed By: thopre

Differential Revision: https://reviews.llvm.org/D85422