LNT: Fix Perf profiling support
In Python3 the C interface has changed. [1]

During the upgrade the Creation of the perf
interface was upgraded to Python3 but the initialization
itself was not. As a consequence the module was not actually
registered and trying to use perf results in an error.

This would not be noticed as the import in the Python module
has the import in a try: .. except: pass block presumably
because perf support is optional.

[1] http://python3porting.com/cextensions.html

Reviewed By: thopre

Differential Revision: https://reviews.llvm.org/D93117