Expect LNT to be installed in the tests
LNT's lit is configured to be able to run the tests without LNT being
installed. This is useful for testing/cPerf.py which skips all its
checks if cPerf extension is not found, which happens when setup has not
run and thus it has not been built.

However newly added runtest/test_suite-profile.shtest expects cPerf
being installed or its globbing fails. This seems a better approach
because a patch author risks missing a regression if some tests are
skipped. Therefore this commit removes the root of the repository from
the PYTHONPATH in lit's config, thereby requiring the user running the
test having lit in PYTHONPATH.

Reviewed By: tnfchris

Differential Revision: https://reviews.llvm.org/D94715