[LNT] Fixed fieldchanges grouping to a regression
Note active_indicators is not a list of regression indicators. It is a list of fieldchanges actually.
After that fix all new fieldchanges will be grouped mostly in the same regression as expected.
It prevents creating of thousand new regressions and ageing them out immediately to keep regressions count under 50.

Reviewed By: thopre

Differential Revision: https://reviews.llvm.org/D109216