[LNT] Fixed possible crash or junk data in cPerf
The return value of the first call of Dump.next() was not checked. Dump.getText() returned an uninitialized string (a junk string) if the objdump did not return an usable data for the specified address range. It caused a crash or inconsistent profile data.

OS Laboratory. Huawei Russian Research Institute. Saint-Petersburg

Reviewed By: thopre

Differential Revision: https://reviews.llvm.org/D112782