[LNT] Fixed incorrect orders behavior in case of miss formatted llvm_project_revision
LNT expected that llvm_project_revision is a version in the format 1.2.3. But it is a string and may contain any symbols. Someone may use hex numbers, dashes, etc. If llvm_project_revision does not contain any digit at all the behavior is undefined and caused inconsistent DB. hash(str) is a workaround to fix this behavior. Note it does not fix the sorting.

Reviewed By: cmatthews

Differential Revision: https://reviews.llvm.org/D109577