lnt: profile queries
Upstream changed LNT a while ago to store the demangled named for C++ functions.

However the demangled names can contain the character , in them, which makes the
split function incorrectly split from the stringified javascript object.

This instead of splitting on , splits on ,[object Object] instead so the lookups
succeed.