Add --filter and --filter-out arguments to test-suite
Sometimes it's useful to ignore a test you know is failing, or to only
run a subset of tests that may reside in different directories. This
adds flags for --filter and --filter-out which just get passed on to
lit.

They're separate from --only-test and --single-result since they don't
interact with the diagnose or predication stuff.