Invoke build tool via cmake --build in test-suite
Instead of directly invoking make, this patch teaches test-suite to call `cmake --build ...` instead, which abstracts away the underlying build tool.

The default CMake generator is still make, but this means that we no longer need to specify the path to the make tool. It also fixes building when the CMake generator is overriden, e.g. with `CMAKE_GENERATOR=Ninja lnt runtest test-suite ...`