Automatically set Ninja generator for --use-make=ninja
When using `--use-make=ninja` or `--use-make=ninja.exe`, the build
would previously fail because CMake was not instructed to generate
the corresponding Ninja build files. It would default to generating
Makefiles for Linux or Visual Studio MSBuild files for Windows.

This change detects when the specified make program is ninja and
automatically adds the `-G Ninja` flag to the CMake config step.