Bump SQLAlchemy==1.3.24 to support lnt on Windows (#33)
LNT crashes on windows on startup with following error in SQLAlchemy:

AttributeError: module 'time' has no attribute 'clock'

The `time.clock` function was removed in Python 3.8, causing this error 
in SQLAlchemy code. The issue was fixed by sqlalchemy/sqlalchemy@003333f

This patch upgrades SQLAlchemy to 1.3.24 which resolves above error.
Version 1.3.24 seemed like a reasonable upgrade containing the required 
fix as it does not contain any backward incompatible changes in some of the 
higher version of SQLAlchemy.