Fix lnt startup crash on Windows (#32)
Running lnt on Windows fails with an Error as it tries to import
the resource module which is not available on windows.

This patch fixes the crash by making the resource module an optional
dependency.

We dont have python module that exposes any function to get the cpu
times of child processes (make.exe + clang-cl.exe + lld.exe). Thus using
wall clock time will be a better alternate.

Also nt TestModule is only used by SPEC when run from llvm-test-suite
and we dont support it on windows.