Handle non-Apple packaged GCC versions
Currently if you use a version of GCC packaged from a Linux distro or built from trunk, you'll get an inferred machine name of something like "foo__gcc_None__riscv64".

The None is supposed to be the cc_build part of the inferred version info, but currently we only handle this for Apple packaged GCC versions.

This teaches the version handling to handle these other types of GCC, which will have a version string like

    gcc version 16.0.0 20250807 (experimental) (GCC)

or

    gcc version 12.2.0 (Debian 12.2.0-14+deb12u1)