Use CMAKE_OBJDUMP when importing profiles in test-suite (#37)
When collecting profiles with `lnt runtest test-suite --use-perf=profile`, it needs a path to objdump to extract the symbols and disassembly. Normally it will try to use `CMAKE_OBJDUMP` from the environment, but AFIAK `lnt` is normally invoked from the command line so no CMake variables are set, so it falls back to just plain old `objdump`.

When running LNT in a cross compilation scenario this leads to a different `objdump` than what CMake uses, which may not support the target (or just straight up not exist). So the profile imports will silently fail to include any code from the binary. 

This fixes it by plumbing through the `CMAKE_OBJDUMP` from the actual llvm-test-suite build. I've left behind the old `os.getenv` as a fallback since `Profile.fromFile` is called from a bunch of other places including `lnt runtest profile`. 

With this patch, using LNT to cross-compile and profile the tests on a remote machine should hopefully "just work"