Run Tox tests in the Github Action CI
This patch enables running the Tox tests in Github Actions CI.
It also adds some dependencies required to run the tests locally
with Tox.

An important change is that we move from LLVM's FileCheck
implementation to a near-equivalent Python version that can be
installed with pip. This will make the test suite more independent
and easier to run: requiring a full LLVM monorepo checkout and a
configured LLVM build just to run the LNT unit tests seems like a
very cumbersome requirement.

Also, this updates the documentation to reflect my current
understanding of how the test suite can be run.

Note that the tests are not passing at the moment, we will need to
update some dependencies and fix other issues. But this gets us
closer to running them, and at least we can see the progress being
made in the CI.