Bump Python requirement to 3.8 and replace usage of pkg_resources (#48)
Before this patch, LNT would use the pkg_resources package to extract
metadata about the current installation. However, that package has been
deprecated in recent Python versions, which produces a warning. Fixing
this either requires switching to the recommended alternative, or
pinning the setuptools dependency to a version < 81.

This patch switches to the recommended alternative (importlib.metadata)
and bumps the Python requirement to 3.8, which is when importlib.metadata
was introduced.

This version bump is fairly gentle and makes us match the Python
requirement used by the rest of LLVM (which is IMO overdue for a
bump as well), so I think it's reasonable.