[tox] Fix a few invalid type annotations
Some type annotations are using nested types of the TestSuiteDB
class. However, these nested types are actually defined in the
__init__ method of TestSuiteDB and are hence not accessible as a
class member. The easiest fix seems to be to downgrade these type
annotations back to comments, at least for now.

In a few other places, add missing annotations to make mypy happy
and add `# noqa` to work around flake8 flagging imports as unused
(when in reality they are required for mypy annotations).

Finally, make the mypy test non-optional so it shows up as a failure
when we break it. We are clearly missing a lot of annotations, but
at least we have a basis that we can build upon and avoid regressing.