Make aggregation function selection consistent across Run and Chart pages (#42)
Currently, the Run page (e.g. /db_default/v4/nts/<runid>) provides a
way to select the aggregation function used when there are multiple
samples in the run (that option is located under "View Options").
The supported options are min and median.

Similarly, the Graph page (e.g. /db_default/v4/nts/graph?<params>)
provides a way to use mean instead of min to aggregate multiple
data points for the same order on the graph. That is provided as a
checkbox under the "Settings 🔧" menu.

This patch makes both selection mechanisms consistent by using a
dropdown menu that provides min, max, mean and median in both cases.

As a drive-by, it reimplements how we select the sample to use as a
reference for related data: we now select the point closest to the
aggregated value instead of selecting the first point unconditionally.