Generate JSON reports in format V2 when using `lnt importreport` (#72)
It seems that a transition towards a new input format had been
started a long time ago, however several parts of the codebase
don't seem to be producing the latest format yet. This moves
`lnt importreport` to the latest format.

As part of the patch, fix capitalization of `Name` for the `Machine`
and the `Test` classes, which didn't match the capitalization
documented and expected by the rest of the code.