Overhaul `lnt updatedb`
That command had a few issues, which I uncovered when I tried updating
the Click dependency. First, it didn't specify the --delete-order
parameter as being a multi-parameter, leading to the default value []
being interpreted as a string with value "[]", i.e. a non-empty string.

This led to the tool always trying to delete an order with value "[]",
which is nonsensical. For some reason, that only happens with newer
Click version.

This patch adds test coverage for --delete-order and makes it a multi
valued parameter, which was obviously always the intent. It also improves
the documentation for the various parameters, which was unclear in an
attempt to be terse.

(cherry picked from commit e4017413f6952fd4c19dbf0873819cc94e5db831)