Bump minimum Python version to 3.10 (#62)
After a recent bump, we documented Python 3.8 as the minimum version.
However, we only tested on Python 3.9 and Python 3.10. In reality, I
get tons of failures with Python 3.9 and only a small number of failures
with Python 3.10.

I think it's makes sense to bump the requirement again if that helps
getting to a stable baseline where CI is passing and our dependencies
are less ancient. In practice, we'll probably deploy the application
using Docker, which means the Python requirement is not critical.