Serialize installation tests
Due to the way installation of the package works (based on setup.py),
it creates temporary content in a build/ directory at the root of the
repository. This sometimes causes installation to fail if another
installation is happening at the same time, which is problematic for
the client/server installation tests.

Until we modernize to a proper pyproject.toml, serialize the installation
tests to ensure they don't fail spuriously.