Rewrite project build and metadata using pyproject.toml
This patch rewrites the project build configuration using pyproject.toml,
which is now the de-facto standard in Python. This single configuration
file supersedes setup.py, tox.ini, mypy.ini, etc.

I checked the contents of the generated package and made sure they matched
exactly (that caught a few issues where I originally excluded some data
files).

This patch is careful not to change the version of dependencies. In
pyproject.toml, we specify minimum package versions for the tool to
work, and we provide a requirements.txt file that pins down specific
versions for use by end-users.

Currently, a few tests are failing if we don't pin exact versions when
running them, so we're still pinning exact versions when we run tests
via tox. The same holds for the type checker.

In the future, we could upgrade towards a proper lockfile instead of
manually constructed requirements.txt files.

Fixes #77