Remove duplicated version and author information (#92)
The information is now canonically stored in pyproject.toml.
It's also worth mentioning that I removed the _version_check() function
because its goal was only to catch misuse of the old `setup.py` "editable"
mode, which we basically don't use anymore for development. We now simply
reinstall the package after making changes, which takes about 2 seconds.