Bump version to 0.4.3
There have been enough changes recently that it probably makes sense
to bump at least the patch version. As a drive-by, I'm removing the
dev0 suffix: this is not a "development" package anymore since it is
being used by various clients and for 10 years or so.

This will allow us to exercise the release process on PyPI.