Only build a source distribution in Github Actions (#98)
Binary distributions for Linux can't be uploaded to PyPI because of ABI incompatibilities.
We could in principle build a binary distribution for macOS, however we don't have many
macOS builders available, which means we'd increase the latency of Github Actions to
potentially hours for the purpose of building a single C++ file.

Instead, just distribute a source distribution.