Don't publish to TestPyPI on pushes to main
It turns out that it was complicated to set up because nobody does it,
and in fact it still doesn't work due to the tagging scheme used by
setuptools-scm. That could be worked around, but since that's not the
intended usage of TestPyPI, don't.

https://github.com/pypa/packaging-problems/issues/844 explains that the
setup guide is misleading since TestPyPI is not intended to be used for
pushes from a CI pipeline.

Hence, only push to TestPyPI on tag creation. This makes me question
the value of TestPyPI in the context of CI/CD, but whatever.