Build the Docker image on pushes, tags, and pull requests
Previously, we'd only build the Docker image on pushes and PRs to main.
In particular, that excluded tags (even on main). With this patch, build
the Docker image all the time, but only push it to ghcr.io when we tag,
which is consistent with how we upload to PyPI.