Get rid of ambiguous negative boolean options with Click
Click 8.3.x introduced a breaking change with respect to how boolean
flags are handled. Indeed, when both a flag_value and a boolean default
are passed, the default value will be assumed to determine whether the
flag is "passed on the command-line".

This differs from the previous behavior, where the default value was
assumed to be the initializer for the flag variable, which is consistent
with how non-boolean values are handled.

Since this is quite a confusing mess, stay far from it by making the
flags more explicit.

This should make the code base work with versions of Click before 8.3.x
and after 8.3.x.