Remove obsolete wsgi_restart configuration option (#122)
I believe this was used at some point to automatically restart the
development server when the LNT sources were changed. However, this
configuration option doesn't appear to be hooked up anywhere, so this
is effectively dead code.

Furthermore, Werzeug supports the `use_reloader=True` argument that
could be used when we start a development server if we wanted this
functionality (which we don't need at the moment).